Rename file to geometry. Remove set incoming beam.
