+++ ReflectometryServer/test_modules/test_components.py
@@ -6,7 +6,7 @@ from parameterized import parameterized
 
 from ReflectometryServer.components import Component, ReflectingComponent, TiltingJaws
 from ReflectometryServer.movement_strategy import LinearMovement
-from ReflectometryServer.gemoetry import Position, PositionAndAngle
+from ReflectometryServer.geometry import Position, PositionAndAngle
 from utils import position_and_angle, position
